fix no Internet connection notification to prevent app from crash, on both admin and client.

+++ DeliveryClient/src/com/mobilive/delivery/client/utilities/RZHelper.java
@@ -39,11 +39,7 @@ public class RZHelper {
 		this.returnMethod = method;
 		this.show = isShow;
 		this.myAQuery = new AQuery(currentActivity);
-		if (!isNetworkAvailable()) {
-			Toast toast = Toast.makeText(myAQuery.getContext(), currentActivity.getString(R.string.no_net),Toast.LENGTH_LONG);
-			toast.setGravity(Gravity.TOP, 0, 0);
-			toast.show();
-		} else {
+		if (checkInternetConnection()) {
 			if (show)
 				loader = new TransparentProgressDialog(currentActivity,R.drawable.spinner);
 
@@ -90,11 +86,7 @@ public class RZHelper {
 		this.last = islast;
 		this.show = false;
 		this.myAQuery = new AQuery(currentActivity);
-		if (!isNetworkAvailable()) {
-			Toast toast = Toast.makeText(myAQuery.getContext(), "Errors: "+ currentActivity.getString(R.string.no_net),Toast.LENGTH_LONG);
-			toast.setGravity(Gravity.TOP, 0, 0);
-			toast.show();
-		} else {
+		if (checkInternetConnection()) {
 			if (first) {
 				loader = new TransparentProgressDialog(currentActivity,
 						R.drawable.spinner);
@@ -139,11 +131,7 @@ public class RZHelper {
 		show = false;
 		myAQuery = new AQuery(currentActivity);
 
-		if (!isNetworkAvailable()) {
-			Toast t = Toast.makeText(myAQuery.getContext(), currentActivity.getString(R.string.no_net),Toast.LENGTH_LONG);
-			t.setGravity(Gravity.TOP, 0, 0);
-			t.show();
-		} else {
+		if (checkInternetConnection()) {
 			myAQuery.ajax(url, File.class, new AjaxCallback<File>() {
 				public void callback(String url, File file, AjaxStatus status) {
 					if (file != null) {
@@ -155,43 +143,57 @@ public class RZHelper {
 	}
 
 	public void post(Object obj) {
-		JSONObject params = (new APIManager()).objToCreate(obj);
-		if (loader != null && show) {
-			myAQuery.progress(loader).post(url, params, JSONObject.class,callBack);
-		}else {
-			myAQuery.post(url, params, JSONObject.class, callBack);
+		if (checkInternetConnection()) {
+			JSONObject params = (new APIManager()).objToCreate(obj);
+			if (loader != null && show) {
+				myAQuery.progress(loader).post(url, params, JSONObject.class,callBack);
+			}else {
+				myAQuery.post(url, params, JSONObject.class, callBack);
+			}
 		}
 	}
 
 	public void get() {
-		if (loader != null && show) {
-			myAQuery.progress(loader).ajax(url, JSONObject.class, callBack);
-		} else {
-			myAQuery.ajax(url, JSONObject.class, callBack);
+		if (checkInternetConnection()) {
+			if (loader != null && show) {
+				myAQuery.progress(loader).ajax(url, JSONObject.class, callBack);
+			} else {
+				myAQuery.ajax(url, JSONObject.class, callBack);
+			}
 		}
 	}
 
 	public void put(Object obj) {
-		JSONObject params = (new APIManager()).objToCreate((Object) obj);
-		if (loader != null && show) {
-			myAQuery.progress(loader).put(url, params, JSONObject.class,callBack);
-		} else {
-			myAQuery.put(url, params, JSONObject.class, callBack);
+		if (checkInternetConnection()) {
+			JSONObject params = (new APIManager()).objToCreate((Object) obj);
+			if (loader != null && show) {
+				myAQuery.progress(loader).put(url, params, JSONObject.class,callBack);
+			} else {
+				myAQuery.put(url, params, JSONObject.class, callBack);
+			}
 		}
 	}
 
 	public void delete() {
-		if (loader != null && show) {
-			myAQuery.progress(loader).delete(url, JSONObject.class, callBack);
-		} else {
-			myAQuery.delete(url, JSONObject.class, callBack);
+		if (checkInternetConnection()) {
+			if (loader != null && show) {
+				myAQuery.progress(loader).delete(url, JSONObject.class, callBack);
+			} else {
+				myAQuery.delete(url, JSONObject.class, callBack);
+			}
 		}
 	}
 
-	private boolean isNetworkAvailable() {
+	private boolean checkInternetConnection() {
 		currentActivity.getApplicationContext();
 		ConnectivityManager connectivityManager = (ConnectivityManager) currentActivity.getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
 		NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
+		
+		if(! (activeNetworkInfo != null && activeNetworkInfo.isConnected())){
+			Toast t = Toast.makeText(myAQuery.getContext(), currentActivity.getString(R.string.no_net),	Toast.LENGTH_LONG);
+			t.setGravity(Gravity.TOP, 0, 0);
+			t.show();
+		}
 		return activeNetworkInfo != null && activeNetworkInfo.isConnected();
 	}
